Extract pruneUnusedJoins.
